ref(relay): Move cabi glob matching into cabi
