+++ relay-cabi/Cargo.toml
@@ -18,6 +18,7 @@ workspace = true
 [dependencies]
 anyhow = { workspace = true, features = ["backtrace"] }
 chrono = { workspace = true }
+globset = { workspace = true }
 json-forensics = { workspace = true }
 lru = { workspace = true }
 once_cell = { workspace = true }
