+++ relay-cabi/src/lib.rs
@@ -102,6 +102,7 @@ mod codeowners;
 mod constants;
 mod core;
 mod ffi;
+mod glob;
 mod processing;
 
 pub use crate::auth::*;
@@ -109,4 +110,5 @@ pub use crate::codeowners::*;
 pub use crate::constants::*;
 pub use crate::core::*;
 pub use crate::ffi::*;
+pub use crate::glob::*;
 pub use crate::processing::*;
