+++ relay-common/Cargo.toml
@@ -15,9 +15,7 @@ workspace = true
 [dependencies]
 chrono = { workspace = true }
 globset = { workspace = true }
-lru = { workspace = true }
 once_cell = { workspace = true }
-parking_lot = { workspace = true }
 regex = { workspace = true }
 sentry-types = { workspace = true }
 serde = { workspace = true }
